refactor(dsl-protobuf): split protobuf model declarations
