+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/ProtobufSchemaType.kt
@@ -1,0 +1,7 @@
+package me.liam.microsmith.dsl.schemas.protobuf
+
+import me.liam.microsmith.dsl.schemas.core.SchemaType
+
+enum class ProtobufSchemaType(override val typeName: String) : SchemaType {
+    PROTOBUF("protobuf"),
+}
