+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/CardinalityField.kt
@@ -1,0 +1,5 @@
+package me.liam.microsmith.dsl.schemas.protobuf.field
+
+sealed interface CardinalityField : Field {
+    val cardinality: Cardinality
+}
