+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/Field.kt
@@ -4,7 +4,3 @@ sealed interface Field {
     val name: String
     val index: Int
 }
-
-sealed interface CardinalityField : Field {
-    val cardinality: Cardinality
-}
