+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/MapType.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.dsl.schemas.protobuf.field
+
+data class MapType(val key: MapKeyType, val value: ValueType) : FieldType
