+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/field/ReferenceField.kt
@@ -1,9 +1,5 @@
 package me.liam.microsmith.dsl.schemas.protobuf.field
 
-import me.liam.microsmith.dsl.schemas.protobuf.types.Type
-
-data class Reference(val name: String, val type: Type? = null) : ValueType
-
 data class ReferenceField(
     override val name: String,
     override val index: Int,
