+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/reserved/Reserved.kt
@@ -9,11 +9,3 @@ sealed interface Reserved {
         }
     }
 }
-
-data class ReservedIndex(val index: Int) : Reserved
-
-data class ReservedToMax(val from: Int) : Reserved
-
-data class ReservedRange(val indexRange: IntRange) : Reserved
-
-data class ReservedName(val name: String) : Reserved
