+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/reserved/ReservedIndex.kt
@@ -1,5 +1,3 @@
 package me.liam.microsmith.dsl.schemas.protobuf.reserved
 
-object Max {
-    internal const val VALUE = 536_870_911
-}
+data class ReservedIndex(val index: Int) : Reserved
