+++ dsl-schemas-protobuf/src/main/kotlin/me/liam/microsmith/dsl/schemas/protobuf/reserved/ReservedRange.kt
@@ -1,0 +1,3 @@
+package me.liam.microsmith.dsl.schemas.protobuf.reserved
+
+data class ReservedRange(val indexRange: IntRange) : Reserved
